Move new tests to the `test` folder
