+++ client-angular/src/app/task-list/deleted/deleted-tasks.component.ts
