--- client-angular/src/app/tasks/active/active-tasks.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ActiveTasksComponent } from './active-tasks.component';
-
-describe('ActiveTasksComponent', () => {
-  let component: ActiveTasksComponent;
-  let fixture: ComponentFixture<ActiveTasksComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ActiveTasksComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ActiveTasksComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
